# executor: apply the `vecGroupChecker` in `window` executor