@Reminiscent 2019-12-03
> Could you add a benchmark result?

Update. PTAL @SunRunAway  @qw4990  @XuHuaiyu